requirements.txt
Django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
webcolors==1.12